training_data/component_render/requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
annotated-types==0.7.0
anthropic==0.45.2
anyio==4.8.0
async-timeout==5.0.1
attrs==25.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
distro==1.9.0
exceptiongroup==1.2.2
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
MouseInfo==0.1.3
multidict==6.1.0
openai==1.63.0
pillow==11.1.0
playwright==1.50.0
propcache==0.2.1
psutil==7.0.0
PyAutoGUI==0.9.54
pydantic==2.10.6
pydantic_core==2.27.2
pyee==12.1.1
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
python3-xlib==0.15
pytweening==1.2.0
requests==2.32.3
sniffio==1.3.1
tenacity==9.0.0
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
yarl==1.18.3
